Added logs, minor changes to db seed

+++ services/web/project/service/service.py
@@ -1,16 +1,21 @@
 """
     The module with buisiness logic
 """
-from ..models import *
-from sqlalchemy import desc
 from math import ceil
-from flask import send_from_directory, current_app
 import os
-from re import match, search
+from re import search
+from sqlalchemy import desc
+from ..models import *
+from flask import session, current_app
+from flask_login import current_user
+from werkzeug.security import generate_password_hash, check_password_hash
 
-film_kwargs_pool = ["name", "release_date", "description", "rating", "poster",
-                   "user_id", "genres", "director"]
-director_kwargs_pool = ["first_name", "last_name", "date_of_birth"]
+film_kwargs_pool        = ["name", "release_date", "description", "rating",
+                           "poster", "user_id", "genres", "director"]
+director_kwargs_pool    = ["first_name", "last_name", "date_of_birth"]
+args_list               = ["page", "limit", "sort", "genre", "director",
+                           "date_from", "date_to"]
+
 
 def get_last_id(model):
     return db.session.query(model).order_by(model.id.desc()).first().id
@@ -27,11 +32,18 @@ def get_item_with_id(item, id):
         return None
 
 
+def get_item_with_filter(data, type, filter):
+    query = type.query.filter(filter == data).first()
+    if query is not None:
+        return "item with {} = {} already exists".format(str(filter).split(".")[1], data)
+    return ""
+
 def delete_item_with_id(serialize, type, id):
     item = get_item_with_id(type, id)
     print(item)
     print(serialize.dump(item))
     if item is None:
+        current_app.logger.warning("No items with id={}".format(id))
         return {"error": "No items with this id"}
     try:
         db.session.delete(item)
@@ -40,16 +52,22 @@ def delete_item_with_id(serialize, type, id):
         print(item)
         return {"deleted": item}
     except:
+        current_app.logger.warning("Something went wrong")
         return {"error": "something went wrong"}
 
 
+def check_author(user):
+    print(user.id, current_user.id, current_user.is_admin)
+    if user.id == current_user.id or current_user.is_admin is True:
+        return True
+    return False
 
 
 def check_director(director_name):
     first_name, last_name = director_name.split(" ")
     if Director.query.where(Director.first_name == first_name).\
             where(Director.last_name == last_name).first() is None:
-        return {"error": "add director first"}
+        return "no director with this name"
     director = Director.query.where(Director.first_name == first_name).\
         where(Director.last_name == last_name).first()
     return director
@@ -66,7 +84,7 @@ def check_film_fileds(**kwargs):
     if "poster" in keys and (search("([A-Za-z0-9])+(\.jpg|\.png)", kwargs["poster"]) is None):
         errors.append("Not a valid filename for poster")
     if "release_date" in keys:
-        if (search("[1-2][0-9]{3}-[0-1][0-9]-[0-3][0-9]", kwargs["release_date"]) is None):
+        if (search("[1-2][0-9]{3}-[0-1][0-9]-[0-3][0-9]$", kwargs["release_date"]) is None):
             errors.append("Valid date format is yyyy-mm-dd")
     if "description" in keys:
         if len(kwargs["description"]) > 1000 or len(kwargs["description"]) < 10:
@@ -119,6 +137,7 @@ def paginate_films(query, **kwargs):
         kwargs["page"] = 1
         query = query.all()
         if len(query) == 0:
+            current_app.logger.warning("No films requested")
             return 0, kwargs
     kwargs["pages"] = ceil(results / kwargs["limit"])
     return query, kwargs
@@ -135,6 +154,7 @@ def sort_films(base_query, **kwargs):
         elif kwargs["sort"] == "rating":
             query = base_query.query.order_by(desc(Film.rating))
         else:
+            current_app.logger.warning("Wrong parameters")
             return "Wrong sort parameter"
     else:
         query = base_query.query.order_by(Film.id)
@@ -154,12 +174,14 @@ def filter_films(query, **kwargs):
         try:
             first_name, last_name = kwargs["director"].split(" ")
         except:
+            current_app.logger.warning("Wrong format for director")
             return {"message": "Enter director if format: <first-name> <last-name>"}
         director = db.session.query(Director).where(Director.first_name == first_name,
                                                     Director.last_name == last_name).first()
         if director is not None:
             query = query.filter(Film.director_id == director.id)
         else:
+            current_app.logger.warning("Wrong director`s name")
             return {"message":"Wrong director`s name"}
     return query
 
@@ -175,12 +197,11 @@ def get_result(query, **kwargs):
 
 
 def get_films(**kwargs):
-    args_list = ["page", "limit", "sort", "genre", "director", "date_from", "date_to"]
     for item in kwargs:
         if item not in args_list:
+            current_app.logger.warning("Wrong parameters")
             return {"message":"Wrong parameters"}
     base_query = Film
-
     #sorting
     query = sort_films(base_query, **kwargs)
 
@@ -196,33 +217,42 @@ def get_films(**kwargs):
 def get_film(id):
     movie = get_item_with_id(Film, id)
     if movie is None:
+        current_app.logger.warning("No films with id={}".format(id))
         return {"error": "No items with whis id"}
     result = film.dump(movie)
     return result
 
 def get_film_img(id):
     movie = get_item_with_id(Film, id)
-    if os.path.isfile("project/static/" + movie.poster) is False:
-        return None
     return movie.poster
 
 
 def add_film(**kwargs):
+    #validate_fields
     errors = check_film_fileds(**kwargs)
+
     if "description" not in kwargs:
         kwargs["description"] = ""
-    if len(errors) > 0:
-        return {"errors": errors}
     if len(kwargs) != 8:
-        return {"errors": "check parameters again"}
+        errors.append("check parameters again")
+
+    check_name = get_item_with_filter(kwargs["name"], Film, Film.name)
+    if len(check_name)>0:
+        errors.append(check_name)
+
+    #check director
     director = check_director(kwargs["director"])
-    if type(director) == type({}):
-        return director
+    if isinstance(director, str):
+        errors.append(director)
+
+    if len(errors) > 0:
+        current_app.logger.warning("errors: {}".format(errors))
+        return {"errors": errors}
+
     user = get_item_with_id(User, kwargs["user_id"])
     genres = get_genres_list(kwargs["genres"])
     id = get_last_id(Film)
     obj = {
-        "id": id+1,
         "name": kwargs["name"],
         "release_date": kwargs["release_date"],
         "description": kwargs["description"],
@@ -235,23 +265,30 @@ def add_film(**kwargs):
     obj.genres.extend(genres)
     db.session.add(obj)
     db.session.commit()
+    current_app.logger.info("Added the film")
     return {"film": film.dump(obj)}
 
 
 def update_film(id, **kwargs):
     errors = check_film_fileds(**kwargs)
-    if len(errors) > 0:
-        return {"errors": errors}
     keys = kwargs.keys()
     if "director" in keys:
         director = check_director(kwargs["director"])
-        if type(director) == type({}):
-            return director
+        if isinstance(director, str):
+            errors.append(director)
         kwargs["director"] = director
     print(kwargs)
     movie = get_item_with_id(Film, id)
+    check_author(movie.user)
+
+    if check_author(movie.user) is False:
+        errors.append("Not an author or admin")
     if movie is None:
-        return {"error": "No items with this id"}
+        errors.append("No items with this id")
+
+    if len(errors) > 0:
+        current_app.logger.warning("errors: {}".format(errors))
+        return {"errors": errors}
     if "genres" in keys:
         genres = get_genres_list(kwargs["genres"])
         movie.genres = genres
@@ -264,27 +301,44 @@ def update_film(id, **kwargs):
 
 
 def delete_film(id):
+    movie = get_item_with_id(Film, id)
+    if check_author(movie.user) is False:
+        current_app.logger.warning("user with id={} is not an author or admin".format(current_user.id))
+        return {"error": "Not an author or admin"}
     return delete_item_with_id(film, Film, id)
 
 
-def get_Directors(page, limit=10):
+def get_directors(page, limit=10):
     results = Director.query.count()
+    if results < 1:
+        current_app.logger.warning("No directors in db")
+        return {"error": "no directors found"}
     result = Director.query.order_by(Director.id).paginate(page=int(page), per_page=limit)
     return {"pages":ceil(results/limit), "page":page, "directors":directors.dump(result)}
 
 
-def get_Director(id):
-    result = director.dump(get_item_with_id(Director, id))
+def get_director(id):
+    result = get_item_with_id(Director, id)
+    if result is None:
+        current_app.logger.warning("No director with id={}".format(id))
+        return {"error": "no directors with this id"}
+    result = director.dump(result)
     return result
 
 
 def create_director(**kwargs):
     errors = check_director_fileds(**kwargs)
+    error = get_item_with_filter(kwargs["first_name"], Director, Director.first_name)
+    if len(error)>0:
+        errors.append(error)
+    error = get_item_with_filter(kwargs["last_name"], Director, Director.last_name)
+    if len(error)>0:
+        errors.append(error)
     if len(errors)>0:
+        current_app.logger.warning("errors: {}".format(errors))
         return {"errors": errors}
-    id = get_last_id(Director)
+
     obj = {
-        "id": id+1,
         "first_name": kwargs["first_name"],
         "last_name": kwargs["last_name"],
         "date_of_birth": kwargs["date_of_birth"]
@@ -298,9 +352,11 @@ def create_director(**kwargs):
 def update_director(id, **kwargs):
     director_update = get_item_with_id(Director, id)
     if director_update is None:
+        current_app.logger.warning("No director with id={}".format(id))
         return {"error": "No items with this id"}
     errors = check_director_fileds(**kwargs)
     if len(errors) > 0:
+        current_app.logger.warning("errors: {}".format(errors))
         return {"errors": errors}
 
     keys = kwargs.keys()
@@ -315,11 +371,52 @@ def delete_director(id):
 
 
 def get_genres():
+    """
+    Returns list of genres from db
+    @return: genres
+    @rtype: list
+    """
     result = Genre.query.all()
+    if result is None:
+        current_app.logger.warning("No genres in database")
+        return []
     genres = [genre.name for genre in result]
     return genres
 
 
-def get_genres_list(genres):
+def get_genres_list(items):
+    """
+    Returns film genres as list
+
+    @param items:
+    @type string:
+    @return: genres
+    @rtype: list
+    """
     return [Genre.query.where(Genre.name == item.lower()).first()
-            for item in genres.split(",")]
+            for item in items.split(",")]
+
+
+def signup(username, password):
+    with current_app.app_context():
+        new_user = User(username=username)
+        new_user.password_hash = password
+
+        db.session.add(new_user)
+        db.session.commit()
+
+        session['user_id'] = new_user.id
+        return new_user.to_dict()
+
+
+def get_user_data():
+    usr = get_item_with_id(User, current_user.id)
+    if usr is not None:
+        return {"user:":models.user.dump(usr)}
+
+
+def set_user_password(password):
+    current_user.password = generate_password_hash(password)
+    usr = get_item_with_id(User, current_user.id)
+    usr.pasword = generate_password_hash(password)
+    db.session.commit()
